IJPL-160982 Refactor condition ExpandRecursivelyAction
To change the condition more easily, we first split it into three
cases: a path is a descendant of a selected path, vice versa,
or otherwise. Now we can add more when branches and modify
existing ones without touching irrelevant code, making the upcoming
changes in the logic easier to perceive.

GitOrigin-RevId: 9ca97d2f04481471f9c8b242b5db878f84a4995f