IJPL-160982 Skip !isIncludedInExpandAll descendants in Expand Selected
Just like in Expand All, if a descendant node doesn't want to be expanded,
we need to skip those here too. An example: those huge Libraries nodes
in the Packages View pane. Nobody ever wants to expand everything
that's in them.

A small difference from the Expand All action: if such a node is actually
selected, and the user presses Expand Selected, we still expand it.
It would make little sense if the action just did nothing in such a case.

GitOrigin-RevId: b02b70b14a81d80d0f24272cb546ed4cfdcc0055