Fix RIDER-100926: IdeDependentAction/Group should update their template presentation to reflect the customized mnemonics
At least the main menu uses the template presentation to determine what action to call by a mnemonic, which leads to confusing results (see the issue description).

Unfortunately, it's not possible to work with the template presentation in the normal way in constructor, because we need action id and thus need access to the action manager, and trying to access the action manager from the action constructor leads to a circular dependency.

So, let's invent a hack that will update the template presentation on the first normal action update.

GitOrigin-RevId: 59930563efac394c3cd225aae4cd698c414fac9c