IJPL-160418 Introduce a unique ID for a Project
To use Project in communication between backend and frontend, we need some sort of serializable unique id that we'd be able to pass through RPC or Rhizome.

In split mode we have to pass this id through RD protocol, because this is how Projects are synchronized there at the moment.

GitOrigin-RevId: c92b1c40830d9f7acf9fd4d0b0e13fa1828e8786