IJPL-160418 Introduce `ProjectEntity`
`Project` is an essential part for most of API.
With Rhizome we need to be able to use `Project` in other entities as well as in RPC.

We can't do that by default, as `Project` is not serializable.
Instead we can introduce `ProjectEntity` which is created for every project in `ProjectManagerImpl`

In additional, util methods `asEntity` and `asProject` are added for convenience

Also, only the backend should be able to delete `ProjectEntity` as it's the one who actually creates it.
The frontend can only delete its `LocalProjectEntity`.
This way we support the situation when a project is closed on the frontend but still available on the backend (e.g., CWM)

GitOrigin-RevId: a0cdefa2ec95930bd32492e9f5261508f0997d4a