IJPL-160418 Fix project leak in tests by enforcing latest change flush
When `ProjectEntity` is deleted with `LocalProjectEntity` all its data stored inside `Novelty` which represents the difference between two DB snapshots (before and after the change).

Therefore, `Project` has an active reference which doesn't allow it to be collected by GC.

`testProjectLeak` checks that all project instances are collected and it fails.

The workaround here is to invoke another change after removing the project entity, which will replace the old `Novelty` with a new one, thus removing the active reference to a project.

GitOrigin-RevId: eafde042b865c503958ec7c2e73a5c056f944ac7