[terminal] IJPL-159812 FUS: Time from command submit to visually started
We need to measure "Time between pressing Enter and showing the active (or already finished) command block."
It is similar to FCP (first contentfull paint) metric.

To achieve this we start timer on "Enter" key pressed and stop when the command is actually sent to the Shell. There could be a delay between because of generators (invisible requests to the Shell) running.
This works because we do not open a new block (visually show that command started) when the Shell reported that command is started, but rather preemptively just after the command is sent to the Shell.

GitOrigin-RevId: f77731785ebaf8e5c6daa87775bf86e557132087