[fleet] rhizomedb: introduce temporary API to migrate lookup and lookupOne using base attributes on subclass
As documented in the code, lookup() was able to work with `subclass::baseAttribute` and return all instances of the subclass that has the base attribute value. With the new API, more boilerplate is necessary at the moment.

For lookupOne, while the replacement would work in all cases, some may benefit from a simple `entity() as? Subclass`, but it has to be decided on a case-by-case basis.

This API would keep track of such cases.

GitOrigin-RevId: be9a435388959a1371e2dd90675f3bb3a6143bec