IJPL-161185 Fix SYS_SCALE for icons cached with OBJ_SCALE
When an icon is scaled, its scale context, normally null,
set to a new instance with OBJ_SCALE corresponding to its
new scaling factor. However, when that context is created,
it also remembers the current system scaling factor.

Later, when the icon is painted, this scale context is used,
and the actual system scaling factor at that time is ignored.
For multi-monitor environments with different monitors
set to different scaling this can lead to the wrong scaling
factor being used. Or when the system scaling factor was
changed after the icon was cached.

Fix by explicitly replacing the cached system scaling factor
with the actual scaling factor of the provided component
or graphics.

GitOrigin-RevId: 82f064a880607eee8997a026829137918151a7ca