[vfs] make VirtualFile.isValid() returns false for 'alien' files, instead of AssertionError
General policy is that alien files must not exist: if VFS is reconnected, _all_ the VirtualFiles from previous session must be thrown out and must not be used _in any way_. By default, we consider _any_ use of such 'alien' files a code bug, so we throw AssertionError if a VFile from a previous VFS epoch is used _in any way_.

Unfortunately,  there are some cases, mainly in older junit3-4 tests, there VirtualFiles _leaked_ from one test to another, with VFS re-connected in between -- which leads to flaky 'Alien file object' assertions failing the tests. So we're forced to compromise: isValid() is the only method that _doesn't_ throw the AssertionError for alien files, but returns false instead.

GitOrigin-RevId: b1f2bb29db32338f73250ec2ce9c9545fcdef1be