IJPL-161185 Implement custom graphics scaling for CachedImageIcon
Now that the bug with the wrong cached system scaling factor is fixed,
some tests started to fail because they "emulate" system scaling
by calling g.scale() on the graphics instance. However, this doesn't
exactly work because the way system scaling is implemented,
it uses the default scaling factor of the graphics configuration,
not the current one of the graphics. For example, if someone
has the system scaling of 1.5, but the painting code called scale(2),
then the total scaling factor will be 3.0. However, JBHiDPIScaledImage
will only handle 1.5. The rest will be accomplished by raster scaling,
which, of course, looks horribly and doesn't match the expected results
of the tests.

To fix that, we sort of emulate the same thing JBHiDPIScaledImage does
by using the object scaling instead and un-scaling the graphics instance
before doing the actual painting. This isn't used in production code much,
but could be used in theory, so we're fixing a potential bug here as well,
not just the tests.

The additional scaling needed is computed using the actual scaling factor
of the graphics and dividing it by the default scaling factor of the graphics
configuration, if any. It's considered to be equal to 1.0 when JRE-side
system scaling is disabled, as in this case JBHiDPIScaledImage doesn't
do its job (it isn't even created), and we have to do everything ourselves
here.

GitOrigin-RevId: c3b1d430b2b86d1710731a94695c932b58446d05