IJPL-60824 Implement the Run Widget minimum size
To actually make the new compressing strategy work,
components inside the toolbar must provide a minimum size.
The Run Widget didn't, so we implement it by calculating
the difference between the full text and the minimum allowed
text.

We'll need this for other widgets as well, so make it an utility method.

GitOrigin-RevId: dc6ff59ffc2bbc90438421383f51b952215c98a4