IJPL-160418 Fix project leak on a new project creation
The flow of a new project creation differs from the flow of an existing project opening.
For a new project, `ProjectImpl` is created quite early and put to `OpenProjectTask` options.

If something goes wrong before we reach the main method `doOpenAsync`, `ProjectImpl` won't be disposed.
The simplest scenario is to cancel project opening when the IDE asks where the project should be opened (controlled by a setting `Settings | System Settings | Open Project In -> Ask`)

If that happens, a newly created project won't be disposed, alongside all preloaded services.

In IJPL-160418 it became noticeable as `ProjectLeakTest.testCloseProject` started to fail. This happens because the ` ProjectImpl ` instance is now held by `LocalProjectEntity` until `ProjectImpl` is disposed

GitOrigin-RevId: 51dba45f9d629ef5dc38a9de13b3b7479e6d86e6