[terminal] IJPL-159812 FUS: Time from command submit to actually started
We need to measure "Time between pressing Enter and receiving the first output from the shell."
It is similar to TTFB (time to first byte) metric.
We actually measure time until Shell reports that the command has started, because the commands could delay text output without us knowing about it. We don't want to measure time of the target executables to print anything.

GitOrigin-RevId: 5ed42f677618e2abe23138bbc2a6779d8b8ab9fb