[kotlin] Remove bytecode inspection for coroutine method switch check
When checking if a suspend method has a state machine, the debugger inspected bytecode trying to find an instruction like `getfield com/example/MainKt$foo$1.label:I`. This approach didn't work in Android Studio, which resulted in a stepping difference between IDEs. This commit introduces a simpler way of detecting if a state machine is generated for a suspend function.