IJPL-161612 Update main toolbar actions on LaF change
Without this, it takes quite a while to update the project widget
icon size. It happens on the next regular update, and those
currently happen twice a second. So the wrong icon gets stuck
until the next update.

We could also update on other events, such as UI settings change,
but it doesn't seem necessary at this point.

GitOrigin-RevId: 83b2894d1acf57dc87291cb088472f3f2bf241c1