[remote dev] don't mark backend and frontend processes as 'isLightEdit'
This is a workaround for IJPL-161632: until a proper condition for setting this flag is implemented, let's explicitly mention that it shouldn't be set if a known command for starting a backend or a frontend is used.

'nosplash' property is now set explicitly for the frontend process to keep previous behavior.

GitOrigin-RevId: 9654279f09ba79d34fdc92660497917f34c2824c