[gitlab] Minimize the data load to send for clones (IJPL-161183)
#IJPL-161183 Fixed

This should make it easier to avoid timeouts,
but we're definitely not guaranteed to prevent them at this moment.

Also splits up the queries for snippets and clone dialog.
These should require and fetch much much less data than full project data.
This is because we may have to fetch hundreds if not thousands
of projects.

GitOrigin-RevId: 06e979df8922533858c117f5cfe2362bb6cb0ab3