GTW-7688 Disable "WelcomeScreen.RevealIn" on the welcome screen
This action is shown under luxed popup so it isn't enough to mark it as `Disabled` - action still going to be shown.

For now, I disable similar actions using `UnattendedHostActionConfigurationCustomizer`
It's the only way to disable the action on the backend side without affecting monolith (e.g. we can't override the actions in `.xml`)

GitOrigin-RevId: 797997064dbf7647a9dae5c9ee489309aa0875ac