[remote dev] don't show the splash screen for the backend process (IJ-CR-144011)
This is a followup to 9654279f09ba79d3. Before, the splash wasn't shown because the backend process was running in 'light edit' mode. After that is fixed, we need to explicitly disable the splash screen for the backend process (it doesn't make sense to show it because user won't see it anyway since the connection isn't established when the splash screen is shown). However, it still makes sense to show the splash screen when running in internal 'splitMode' when both backend and frontend processes run on the same machine, to make behavior similar to behavior in the monolith mode.

GitOrigin-RevId: 739eea36ec9b9c4eb15ea7bd195e80c0c340a1cd