[terminal] Send commandEndMarker *after* command_finished event OSC.
The codebase is built on the assumption that the marker is going after the event at the very end of the output. This is not broken for now because of the coincidence that event osc is invisible for output and marker becomes the last output.
Send the marker after the event to make it correct and stable.

GitOrigin-RevId: 8d2e4e93556f5995a635026e80263e6e20fff2e8