[modular loader] refactoring: rename ModuleImportance and its entries for consistency (IJPL-128)
After IJPL-148565 is fixed, we now have ModuleLoadingRule enum for modules included in plugins. Since the meaning of its entries is the same as the meaning of ModuleImportance's entries, the enum and its entries are renamed for consistency. Currently, we cannot reuse RuntimeModuleLoadingRule for plugins to avoid adding dependency from intellij.platform.core.impl to intellij.platform.runtime.product module.

GitOrigin-RevId: c508919804af4a1910c4ee1ce2253228c4164b82