IJPL-453 re-throw logged errors from the test thread right after the test
Before the change, `Logger.error` calls threw errors.
After this change, errors, which are reported in the thread running the test,
are collected and then re-thrown right after the test.

Errors, which are reported on other threads, are re-thrown as usual.

This is the first step towards making `Logger.error` non-throwing.

GitOrigin-RevId: aeb24b48a29f2349f6c03c91117df05aaa43a9ff