IJPL-159930 Do not insert a lot of line breaks in case when part of the output is discarded from the text buffer
In cases when the output is very big, and we insert it in the document much slower when it arrives to the text buffer, we receive the partial output that starts on some line far out of the current document bounds. Previously we inserted all missing lines and then they were trimmed. It was not really effective.


(cherry picked from commit bbec3b14754970ef39fe65abecc046091e20f340)

IJ-MR-144634

GitOrigin-RevId: 420add7faac1392a6bd1b90e20bb509d5f659712