Python: Add methods to bind flows to Kotlin DSL components.
Thanks to `Component.showingScope` we can now collect data from a flow as long as the component is displayed.

At the same time, we can emit data back using `MutableStateFlow`.

We can now move UI-specific logic to the View Model and cover it with tests, leaving a View as simple as possible.

GitOrigin-RevId: 7df699d1428506e8c42cffebbb31702fd248e004