Cache included files (along with their user data) by soft reference (part of KTIJ-27513 K2 IDE: Slow reference search for specific symbols in unused symbol inspection)
During the highlighting, resolve happens which needs a scope. To compute the scope, the project structure is needed, to compute which the plugin dependencies structure is needed, to compute which the plugin.xml needs to be analyzed along with all its includes. These includes are cached in the user data of the included file, which is retained by weakref in FileManagerImpl, meaning it's gced very often, clearing all its caches in the process. To make these caches live a bit longer, retain all these included files by soft ref in FileIncludeManager

GitOrigin-RevId: 6fc22d23af0e82756b34173a8e03680c46cfb899