IJPL-159930 Refactor: use the same logic of trimming the last empty line as it was before rework
But now it is performed right in the editor.
Trimming can't be performed on the `PartialCommandOutput` content because in the case with empty command output, we anyway need to remove the line break that we add after the command text.
Also, move the comment with a detailed description.


(cherry picked from commit dee0b7261dfee936cc4fb437835e23ca33ecd991)

IJ-MR-144634

GitOrigin-RevId: 8adedc6085532e3414150014ea1590096a4fc8b2