[json] IJPL-160634 Fix accidentally broken json schema mapping persisting
Historically, we store enum's string representation inside the json mappings state. In turn, the string is assembled from a message bundle line used in several places across the project. Changing this line broke the deserialization process.

To preserve backward compatibility, we intentionally continue using the same approach, but provide a converter capable of reading the schema version tag both in new and in old formats

GitOrigin-RevId: 8813740318ede3c9bc5656ee46c847059760e3c5