IJPL-159930 Wait for initial block finalization in tests
After the rework of document update, we have to rely on the fact that command started event is handled on EDT (install content listener and so on) before the command finished event is arrived.
Without this fix `command with empty output` test is failing for Zsh.
The general problem of event sequence guarantees is important and will be addressed later.


(cherry picked from commit 0bcbc6a7e7741be1cfc2443c76e5f58a6c393dda)

IJ-MR-144634

GitOrigin-RevId: 1a723219b5f349a153506eae7e964c5da5faa3bd