[java] Don't add module dependency in an undo-transparent action
This action should be undoable in most cases, and some implementors of the JavaProjectModelModifier execute the actual change inside a command.
In particular, it fixes AMPER-2132.

This was an accidental change in d634f0d.

However, since all java project modifications should be done within a command, including the optional jigsaw module update, the latter should be wrapped into an undoTransparent command, to be undoable together with the main add module dependency action.

GitOrigin-RevId: 3c0c3b5ce4747c904a291385bcedea813be76892