IJPL-158493 Introduce ProjectFileNodeUpdaterInvoker
To reuse ProjectFileNodeUpdater for the new coroutine-based
Project View, we need to remove its direct dependency on the legacy
Invoker. Introduce a new abstraction layer, including only
the methods that are really needed.

Because the existing implementation relies on the invoker
as the parent Disposable, we must implement it in the new
interface as well, and ensure that our invoker is disposed
along with the delegate.

To remove the dependency in AsyncProjectViewSupport,
remove the invoke() call around roots.forEach. It's a safe thing to do
because the ProjectFileNodeUpdater itself only ever invokes
updateStructure in the same invoker, so that wrapping
was effectively a no-op.

GitOrigin-RevId: dd239b1330a2d8ef84f09e1dc5e3a49d79d8414c