IJPL-158493 Extract a superclass from AsyncProjectViewSupport
To avoid duplicating a lot of logic in the new coroutine-based
implementation, extract the parts of AsyncProjectViewSupport
that don't depend on the models into a superclass.

Refactor the code around acceptAndUpdate a bit so
the only model-dependent part is now in this function
with a new signature, so it can be used in all places
where a similar logic was used.

Other than that, the code was mostly just moved.
The part about myNodeUpdater is a bit ugly,
as it's now initialized in the derived class and then
used in the base class in setupListeners.
It's OK, as it's an internal implementation class,
it won't have a lot of subclasses, so we're not forcing
anyone to figure out this ugliness before using it.

GitOrigin-RevId: a8d858c5c579a538a4c917cf830bc0704468f43d