IJPL-158493 Add new tree model support to Project View
Implement ProjectFileNodeUpdaterCoroutineInvoker and
CoroutineProjectViewSupport to bridge the existing implementation
of the Project View and the new tree model implementation.

Add the ide.project.view.coroutines registry key to be able
to enable the new implementation, but keep it disabled for now,
as the new implementation likely needs a lot of fixes and improvements.

GitOrigin-RevId: 532f1c18a0a48f34d4d07c474e9345f9ba4023ca