IJPL-161797 Optimize calculateComponentWidths
No need to ++ by a single pixel on every iteration.
If there are other components left, we should at least increase
the component's width so it changes its position in the queue,
and we take another component on the next iteration.
Without exceeding its preferred width, of course.

If it's the last component, just extend it to its preferred size.

In both cases take care to not exceed the actually available
width.

GitOrigin-RevId: 73476a870e28873a761cdd6f38ee9b7881a4947d