[remote dev] don't fail if the backend process is started with an unapproved third-party plugin (GTW-9680)
A dialog asking a user to give consent to use a third-party plugin cannot be shown in remote dev mode because the platform isn't fully loaded at that moment, and the connection is not initialized yet. So to avoid exceptions, now in such cases third-party plugins are automatically disabled, and a warning is printed to the log with a suggestion about using a new '--give-consent-to-use-third-party-plugins' option in the 'installPlugins' command (see IJPL-162211).

GitOrigin-RevId: 62907bd7c9e70e37278f3699c550852ed0790a23