IJPL-158493 Fix Select in Project View for the coroutine implementation
We must visit the Swing model there because we don't just do it to collect
paths to select. We also must make sure that the collected paths become
known to the Swing model, because it'll be impossible to actually select
them otherwise.

Alternatively, we could collect the paths first, and then perform selecting
and expanding using another visitor, which would be visiting the Swing
model, but it would be a more complicated solution with no added benefits.
We're already visiting the model, so it's best to do everything in one go
and then just let Swing to do its job without any more background stuff.

GitOrigin-RevId: fc4d2871d92fbc1cc0316c06f7faa41cf27e225f