[plugins] provide an option to automatically approve third-party plugin in the 'installPlugins' command (IJPL-162211)
Before, if 'installPlugins' commands installs a third-party plugin, it recorded its ID in a file, and the next time the IDE is started, it asks a user for confirmation. It was inconvenient because if an IDE installation is prepared by scripts, it isn't expected that the user needs to confirm something manually. Also, the confirmation window is shown during an early phase of the startup, and it cannot be shown to the user in remote development mode (see GTW-9680).

So now a new option '--give-consent-to-use-third-party-plugins' for the 'installPlugins' command is supported. If it's specified, it automatically marks installed third-party plugins as approved.

GitOrigin-RevId: 812ce8be018b722a5c676fb879234ef1e3e80bdb