[platform] Remove LocalProjectEntity, use only ProjectId and ProjectEntity
The main issue in LocalProjectEntity is its creation. LocalProjectEntity is a local entity that depends on shared ProjectEntity. There is a race in shared entity creation, it may be created but then overwritten during rebase, but LocalProjectEntity may depend on the older version of ProjectEntity.
So, for now we will just remove LocalProjectEntity

GitOrigin-RevId: 77d69fb231d80147d11699875ce9ab810a72b965