IJPL-160386 Move task subscription to a different service
Each task has a project entity which indicates the frame where the progress should be shown.
Before, all tasks were collected in `PlatformTaskSupport` application service.
However, it required waiting for a project if `TaskInfoEntity` arrived too early (before a project is created on frontend)

Without `LocalProjectEntity` (which was removed recently) waiting for a project becomes more complicated.
Instead, it'd be better to have a separate per-project service which listens for tasks for a provided project only.

GitOrigin-RevId: 36a486a507da98d707c7f85cf2dabc0fd826f800