IJPL-162180 Make ToolWindowStateCollector more thread-safe
Don't access the windowsSize map twice to check the old value.

Use ConcurrentHashMap for collectors, as the tool window
manager isn't guaranteed to always fire events on the EDT.
This is especially true for the "unregister" events
that we're about to start handling.

Replace getOrPut with computeIfAbsent, as getOrPut
doesn't guarantee that the supplied function isn't called
if the map already contains the key.

GitOrigin-RevId: bfec867346030603aef8729e96afe2e74cc679b2