IJPL-159968 Move disableLayoutInTextComponents and co to another class
Because it's used in initLux, it's accessed at an early delicate stage.
But UiUtil is a huge class that invokes a lot of stuff in its static init,
and some of that stuff (JBInsets.create) involves scaling, which isn't
initialized here yet.

Invoking preload() from there was a mistake. It's not guaranteed
that all data needed to compute the scaling factors is ready.
And we don't really need it at that stage.

To fix this properly, we extract just the part of UiUtil that's
used in initLux. It isn't even a public API. And as a bonus,
it'll be faster here too, as we don't need to load that huge class
anymore.


(cherry picked from commit 6ad6bbb24c7753cfb27f148e405a31117ad74964)

IJ-CR-142474

GitOrigin-RevId: 4d329975ab3f3f045352e54e45b9de2090175549