IJPL-159968 Introduce a local variable in patchLafFonts
Since this is a computable property, to be on the safe side,
read it only once. This can be very helpful if ever in the future
it starts returning different values due to some bug,
for example. Then we can be sure that the used value and
the logged value are the same.


(cherry picked from commit 0ca8e953c353ea4405f0d189399f11ea08a9aacf)

IJ-CR-142474

GitOrigin-RevId: b610dd9a6f41e33c541bf14334364ad79ad420e2