[remote dev] refactor representation of downloaded frontend
Before we used ExtractedJetBrainsClientData to represent both standalone separate JetBrains Client distributions and a distribution embedded in a monolith IDE. This may cause confusion because the frontend process is started differently in these cases. So to simplify code and make it easier to migrate to the new launcher for JetBrains Client (GTW-9619), a sealed 'FrontendInstallation' interface with separate implementations is introduced.

(Changes in spaceport were reviewed in IJ-MR-145468.)

GitOrigin-RevId: c2355c517b65bfaf4e34b5c4b25cd55870a7ec58