[vcs] [RD] IJPL-160237 Added FileStatus support
Since shelves use only a restricted list of file statuses that are already known (they can't be contributed by plugins in this case), so we can just pass an ID. The static map is used as a temporary solution because we can't get an instance of FileStatusFactory and ask it for all our statuses as they can be not initialized yet. So, we either need to initialize and force the FileStatus interface loading with some workarounds or use reflection.

GitOrigin-RevId: d4cc977edd344edeeb86d2c66b18602d72f0728e