Save all documents when compiling the project before running a configuration
Before this fix, if a run configuration delegated to an external task to build the project, it was the responsibility of the corresponding ProjectTaskRunner to save documents before running. It led to situations when each ProjectTaskRunner implementation should remember to save before build (and some forget to do it: AMPER-2958). However, it is more reasonable for the platform to do it once and for all.

GitOrigin-RevId: 87c460ecb2160fc36ed557d9faa3ef1ff28c1c80