[json] IJ-CR-144628 Redistribute split module inside end-user plugin jars
If we could specify custom path for required content modules, we could include json.split into the monolith intellij.json plugin as well.

Currently we have to keep it in a separate jar to satisfy the test. It seems to make litle sense since no other plugin in monolith ide depends on json.split.

GitOrigin-RevId: 77362ce4de0957d941844003d69593c273269a68