Fix encoding of project name in doc url
JavaDocExternalFilter.getExternalDocInfoForElement does a comparison to
determine if the given doc URL is coming from the built in server.
However, it does this using a constructed string that fails to correctly
encode the project name, so the comparison does not work for any project
names that contains characters needing encoding (like a space).

Using Urls.newHttpUrl() to make the comparison string fixes the issue,
since this is the same API used in BuiltInWebBrowserUrlProvider to
create the doc URL in the first place.

This change matches exactly what I've proposed for the platform with
https://github.com/JetBrains/intellij-community/pull/2851.

Fixes: 369323271
Test: Manually verified
Change-Id: If255eb6c8a6633b4eaee436f1dce344c4c8ec12c