IDEA-381622 javadoc: fix encoding of project name in doc url
JavaDocExternalFilter.getExternalDocInfoForElement does a comparison to
determine if the given doc URL is coming from the built in server.
However, it does this using a constructed string that fails to correctly
encode the project name, so the comparison does not work for any project
names that contains characters needing encoding (like a space).

Using Urls.newHttpUrl() to make the comparison string fixes the issue,
since this is the same API used in BuiltInWebBrowserUrlProvider to
create the doc URL in the first place.

closes https://github.com/JetBrains/intellij-community/pull/2851
#IDEA-381622 Fixed

GitOrigin-RevId: 54b9a4c40e0a265426fdcf09ff192528f9a48428