Fix z-order for text field/area placeholder (#615)
This changes the order in which the placeholder and actual text field
are laid out in the TextField and TextArea code. Since the two have the
same zIndex (the default value, 0f) their placement order is reflected
in the drawing order: the last placed is drawn last, on top of the rest.

This fixes a minor cosmetic issue, where the placeholder would be drawn
on top of the cursor when the TextArea/TextField is empty.
GitOrigin-RevId: 1238a6dc8147323ea0ce1f80bc89e808f27a3663