Fix TextArea placeholder layout (#616)
Fix TextArea placeholder layout

When we added scrollbars and tweaked the TextArea paddings, we did not
do the same for the placeholder. As a result, the placeholders were
laid out incorrectly, up against the component's border. This fixes that
issue by passing down the modifier to the TextArea decoration box.
GitOrigin-RevId: 1cd93c4247e48c11973228b7f3a808c6c5abeb8f