[remote dev] move the rhino library back from lib.jar to lib-client.jar (GTW-9946)
'intellij.platform.ide.util.io.impl' module depends on 'proxy-vole' library, where transitive dependency on rhino-runtime is excluded. So we need to have explicit dependency on 'rhino' library there. Otherwise, it is included in lib.jar instead of lib-client.jar, and isn't included in the classpath of the frontend process started from the IDE, leading to NoClassDefFoundError errors.

Before it worked because 'intellij.platform.ide.impl' had dependency on 'rhino' library, but it was removed in 124621881da764.

GitOrigin-RevId: 221684b33a90b1be65c6c4d48b3a64417db7c9b1