[inline-completion] mark `InlineCompletionEvent` as non-extendable
It becomes hard to evolve the Inline Completion feature when there are unknown implementations of such basic interfaces. If you are missing a built-in inline completion event, please create a feature request for it.

It's not completely forbidden to create custom inline completion events. But be aware that some built-in features might not work with them.

GitOrigin-RevId: c2097d10d74910326e8f50e13dc52bd13f2b488e