IJPL-160520 Log debug info about JBPopupMenu placement
JPopupMenu invokes adjustPopupLocationToFitScreen()
to make the menu fit into the screen. It's quite complicated,
and to be able to figure out why the menu showed up in
a specific place, we need to know everything that method uses,
including its input parameters (what's passed to show()),
the invoker component, information about the screens and their insets.

For TRACE let's log the stack too, as it may help us to understand
where the input values came from.

GitOrigin-RevId: 0df4bdfe6f845af919917c4031105d0a2881d039