[RDCT] IJPL-162126: Fix availability hints exceptions
`EditorActionAvailabilityHintHandler` adds hints on every highlighter update. If there is a highlighter with action availability hints and subsequent updates, these hints will be added multiple times, causing `logger.error`. The same happens in some other places like `InlinePromptHighlightingPassFactory` which may add hints several times.
This commit changes the assertion from "log error if there is a hint already" to "log error only if there is a *different* hint for this action". It skips already existing hints if they are exactly the same.

GitOrigin-RevId: 83e88ba3a050e291b5634341dd7526620623858f